requirement.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1
djangorestframework==3.15.2
executing==2.1.0
gunicorn==23.0.0
ipython==8.27.0
jedi==0.19.1
kombu==5.4.0
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
